Rename _deprecation module to deprecation
