+++ src/repobee_plug/__init__.py
@@ -25,7 +25,7 @@ from repobee_plug._exthooks import SetupHook as _setup_hook
 from repobee_plug._exthooks import ConfigHook as _config_hook
 
 # Helpers
-from repobee_plug._deprecation import deprecate, deprecated_hooks, Deprecation
+from repobee_plug.deprecation import deprecate, deprecated_hooks, Deprecation
 from repobee_plug._serialize import (
     json_to_result_mapping,
     result_mapping_to_json,
