+++ tests/unit_tests/repobee_plug/test_deprecation.py
@@ -1,7 +1,7 @@
 import pytest
 
 import repobee_plug as plug
-from repobee_plug import _deprecation
+from repobee_plug import deprecation
 
 
 def test_cant_deprecate_non_hook_function():
@@ -9,7 +9,7 @@ def test_cant_deprecate_non_hook_function():
         pass
 
     with pytest.raises(plug.PlugError) as exc_info:
-        _deprecation.deprecate(remove_by_version="3.0.0")(func)
+        deprecation.deprecate(remove_by_version="3.0.0")(func)
 
     assert "can't deprecate non-hook function" in str(exc_info.value)
     assert "func={}".format(func) in str(exc_info.value)
